requirements.txt
openvino-dev[tensorflow2]
matplotlib
